events/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
jinja2==3.1.3
aiofiles==23.2.1
httpx==0.26.0
redis==5.0.1
celery==5.3.6
pypdf2==3.0.1
reportlab==4.0.9
weasyprint==60.2
pillow==10.2.0
pyyaml==6.0.1
python-dotenv==1.0.0
sendgrid==6.11.0
icalendar==5.0.11
hcaptcha==0.1.0
